+++ HabboHotel/Items/Interactor/InteractorWired.cs
@@ -25,7 +25,7 @@ public class InteractorWired : IFurniInteractor
         item.LegacyDataString = "1";
         item.UpdateState(false, true);
         item.RequestUpdate(2, true);
-        if (item.Definition.WiredType == WiredBoxType.AddonRandomEffect)
+        if (item.Definition.IsRandomWiredAddon)
             return;
         if (item.GetRoom().GetWired().IsTrigger(item))
         {